jenkins-examples - Update Tool-Publish-* to use publish-bknix environment

+++ jenkins-examples/Tool-Publish-amp.bash
@@ -1,21 +1,22 @@
 #!/bin/bash
 set -ex
 
+#### Setup environment
+if [ -e "$HOME/.profile" ]; then . "$HOME/.profile"; fi
+eval $(use-bknix min)
+if [ -z "$BKITBLD" ]; then echo "Invalid BKPROF"; exit 1; fi
+bknix update
+
 #### Config
 PROGRAM=amp.phar
 TMPFILE=amp.phar
 REVISION=$(date +'%Y-%m-%d')-$(git rev-parse HEAD | head -c8)
-# PHP=/opt/php/5.4.45/bin/php
-PHP=/usr/bin/php
-
-#export PATH="$HOME/buildkit/bin:$HOME/bin:$PATH"
-export PATH="/srv/buildkit/bin:$HOME/bin:$PATH"
 
 #### Build it
 [ -d vendor ] && rm -rf vendor
 [ -f "$TMPFILE" ] && rm -f "$TMPFILE"
-$PHP `which composer` install --no-scripts
-$PHP -dphar.readonly=0 `which box` build
+composer install --no-scripts
+php -dphar.readonly=0 `which box` build
 
 #### Publish
 [ -d tmp ] && rm -rf tmp
